requirements.txt
# ============================================================================
# DEPENDENCIAS ESENCIALES - Chatbot RAG TechStore Perú
# ============================================================================

# Framework Web
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
jinja2>=3.1.4
python-multipart>=0.0.9

# Base de Datos
sqlalchemy>=2.0.32
PyMySQL>=1.1.1

# Seguridad
passlib[bcrypt]>=1.7.4
itsdangerous>=2.2.0

# Configuración
python-dotenv>=1.0.1

# ============================================================================
# RAG (Retrieval-Augmented Generation)
# ============================================================================

# Framework de orquestación
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1

# Vector Database
chromadb==0.4.24

# Embeddings
sentence-transformers==2.5.1

# Procesamiento de documentos
PyMuPDF==1.23.26        # PDFs
python-docx==1.1.0      # Word
pandas==2.1.4           # Excel
openpyxl==3.1.2         # Excel backend

# Chunking y text processing
tiktoken==0.6.0

# Utilidades
filetype==1.2.0
